refactor(storage): extract variations + result queries from sqlite_store
A4 of the sqlite_store decomposition. Two extracts:

- amx/storage/_history_results.py: get_run_results, get_run_result,
  list_runs_with_result_counts. Read-only queries used on every run
  detail / list page.
- amx/storage/_history_variations.py: 8 methods that drive the
  variation tree + rerun-snapshot lifecycle - get_descendant_runs
  (178 LOC), get_result_chain, next_rerun_seq, save/read/list/delete
  rerun_snapshot, gc_orphan_rerun_snapshots.

Each function takes the store as 'hs' and uses _connect(). No DDL
touched. SQLiteHistoryStore keeps thin delegating wrappers on every
name. Existing test coverage (test_variations_tree, test_rerun_chain,
test_get_descendant_runs) exercises this path.

sqlite_store.py: 2480 -> 2049 LOC (-431, -17%). Cumulative across
A1+A2+A3+A4: 3535 -> 2049 LOC (-42%) over four PRs.
